# frontend-test-go-server/api/swagger.yaml
...
swagger: "2.0"
info:
  description: "The registry API for SwaggerHub"
  version: "1.0.97"
  title: "SwaggerHub Registry API"
  contact:
    name: "SwaggerHub"
    url: "http://swaggerhub.com"
    email: "info@swaggerhub.com"
host: "dev-api.swaggerhub.com"
...
schemes:
- "https"
produces:
- "application/json"
paths:
...
  /domains:
    get:
      tags:
      - "Domains"
      summary: "Retrieves a list of currently defined domains in APIs.json format"
      operationId: "searchDomains"
      parameters:
      - name: "filter"
        in: "query"
        description: "filter to apply, 'user' filters by current user"
        required: false
        type: "string"
        enum:
        - "user"
        x-inflector-hidden: true
        x-exportParamName: "Filter"
        x-optionalDataType: "String"
      - name: "query"
        in: "query"
        description: "free text query to match"
        required: false
        type: "string"
        x-exportParamName: "Query"
        x-optionalDataType: "String"
      - name: "state"
        in: "query"
        description: "matches against published state of the spec\n* UNPUBLISHED -\
          \ spec is a draft, a work in progress\n* PUBLISHED - spec is a stable version\
          \ ready for consuming from client applications\n* ANY - either PUBLISHED\
          \ or UNPUBLISHED\n"
        required: false
        type: "string"
        default: "ALL"
        enum:
        - "ALL"
        - "PUBLISHED"
        - "UNPUBLISHED"
        x-exportParamName: "State"
        x-optionalDataType: "String"
      - name: "tag"
        in: "query"
        description: "matches against tags associated with a domain"
        required: false
        type: "array"
        items:
          type: "string"
        collectionFormat: "multi"
        x-exportParamName: "Tag"
      - name: "page"
        in: "query"
        description: "page to return"
        required: false
        type: "integer"
        default: 0
        format: "int32"
        x-exportParamName: "Page"
        x-optionalDataType: "Int32"
      - name: "limit"
        in: "query"
        description: "number of results per page (1 .. 100)"
        required: false
        type: "integer"
        default: 10
        format: "int32"
        x-exportParamName: "Limit"
        x-optionalDataType: "Int32"
      - name: "sort"
        in: "query"
        description: "sort criteria or result set\n* NAME -\n* UPATED\n* CREATED\n\
          * OWNER\n* BEST_MATCH\n* TITLE\n"
        required: false
        type: "string"
        default: "NAME"
        enum:
        - "NAME"
        - "UPDATED"
        - "CREATED"
        - "OWNER"
        - "BEST_MATCH"
        - "TITLE"
        x-exportParamName: "Sort"
        x-optionalDataType: "String"
      - name: "order"
        in: "query"
        description: "sort order"
        required: false
        type: "string"
        default: "ASC"
        enum:
        - "ASC"
        - "DESC"
        x-exportParamName: "Order"
        x-optionalDataType: "String"
      responses:
        "303":
          description: "303 response"
          headers:
            Location:
              type: "string"
      security:
      - TokenSecured: []
      - {}
      x-swagger-router-controller: "com.smartbear.swaggerhub.registry.service.DomainsV1Service"
  /domains/{owner}:
    get:
      tags:
      - "Domains"
      summary: "Retrieves an APIs.json listing of all domains defined for this owner"
      operationId: "getOwnerDomains"
      parameters:
      - name: "owner"
        in: "path"
        description: "API or Domain owner identifier (case-sensitive)"
        required: true
        type: "string"
        x-exportParamName: "Owner"
      - name: "page"
        in: "query"
        description: "page to return"
        required: false
        type: "integer"
        default: 0
        format: "int32"
        x-exportParamName: "Page"
        x-optionalDataType: "Int32"
      - name: "limit"
        in: "query"
        description: "number of results per page (1 .. 100)"
        required: false
        type: "integer"
        default: 10
        format: "int32"
        x-exportParamName: "Limit"
        x-optionalDataType: "Int32"
      - name: "sort"
        in: "query"
        description: "sort criteria or result set\n* NAME -\n* UPATED\n* CREATED\n\
          * OWNER\n* BEST_MATCH\n* TITLE\n"
        required: false
        type: "string"
        default: "NAME"
        enum:
        - "NAME"
        - "UPDATED"
        - "CREATED"
        - "OWNER"
        - "BEST_MATCH"
        - "TITLE"
        x-exportParamName: "Sort"
        x-optionalDataType: "String"
      - name: "order"
        in: "query"
        description: "sort order"
        required: false
        type: "string"
        default: "ASC"
        enum:
        - "ASC"
        - "DESC"
        x-exportParamName: "Order"
        x-optionalDataType: "String"
      responses:
        "200":
          description: "a list of domains in APIs.json format"
          schema:
            $ref: "#/definitions/ApisJson"
      security:
      - TokenSecured: []
      - {}
      x-swagger-router-controller: "com.smartbear.swaggerhub.registry.service.DomainsV1Service"
  /domains/{owner}/.refs:
    get:
      tags:
      - "Domains"
      summary: "Retrieves an APIs.json listing of entries referencing owner domains"
      operationId: "getOwnerReferences"
      parameters:
      - name: "owner"
        in: "path"
        description: "API or Domain owner identifier (case-sensitive)"
        required: true
        type: "string"
        x-exportParamName: "Owner"
      responses:
        "200":
          description: "a list of APIs and domains in APIs.json format"
          schema:
            $ref: "#/definitions/ApisJson"
      security:
      - TokenSecured: []
      x-inflector-hidden: true
      x-swagger-router-controller: "com.smartbear.swaggerhub.registry.service.DomainsV1Service"
  /domains/{owner}/.suggest/{type}:
    get:
      tags:
      - "Domains"
      summary: "Suggests domain entites of the specified type by their name"
      operationId: "suggestDomainEntity"
      parameters:
      - name: "owner"
        in: "path"
        description: "API or Domain owner identifier (case-sensitive)"
        required: true
        type: "string"
        x-exportParamName: "Owner"
      - name: "type"
        in: "path"
        description: "entity type"
        required: true
        type: "string"
        enum:
        - "pathitems"
        - "pathItems"
        - "definitions"
        - "parameters"
        - "responses"
        - "schemas"
        - "examples"
        - "requestbodies"
        - "requestBodies"
        - "headers"
        - "links"
        - "callbacks"
        x-exportParamName: "Type_"
      - name: "query"
        in: "query"
        description: "query string"
        required: false
        type: "string"
        x-exportParamName: "Query"
        x-optionalDataType: "String"
      - name: "oas"
        in: "query"
        description: "The OpenApi Specification (oas) version.\n* 2.0 - Formally known\
          \ as 'swagger'\n* 3.0.x - The latest supported OpenApi Specification version\
          \ by SwaggerHub.\nFor search queries all 3.0.x versions are grouped together\n"
        required: false
        type: "string"
        enum:
        - "2.0"
        - "3.0.0"
        - "3.0.1"
        - "3.0.2"
        - "3.0.3"
        x-exportParamName: "Oas"
        x-optionalDataType: "String"
      responses:
        "200":
          description: "a list of domains in APIs.json format. Each entry will be\
            \ listed with the 'X-Suggestion' property  which value will contain comma\
            \ separated list of available definitions matching the query"
          schema:
            $ref: "#/definitions/ApisJson"
        "404":
          description: "specified owner not found"
      security:
      - TokenSecured: []
      - {}
      x-inflector-hidden: true
      x-swagger-router-controller: "com.smartbear.swaggerhub.registry.service.DomainsV1Service"
  /domains/.suggest/{type}:
    get:
      tags:
      - "Domains"
      summary: "Suggests all domain entites of the specified type. This endpoint returns\
        \ any entities that the user has access (either as an API owner, Org Member\
        \ or Collaborator)."
      operationId: "getUserDomains"
      parameters:
      - name: "type"
        in: "path"
        description: "entity type"
        required: true
        type: "string"
        enum:
        - "pathitems"
        - "pathItems"
        - "definitions"
        - "parameters"
        - "responses"
        - "schemas"
        - "examples"
        - "requestbodies"
        - "requestBodies"
        - "headers"
        - "links"
        - "callbacks"
        x-exportParamName: "Type_"
      - name: "query"
        in: "query"
        description: "query string"
        required: false
        type: "string"
        x-exportParamName: "Query"
        x-optionalDataType: "String"
      - name: "oas"
        in: "query"
        description: "The OpenApi Specification (oas) version.\n* 2.0 - Formally known\
          \ as 'swagger'\n* 3.0.x - The latest supported OpenApi Specification version\
          \ by SwaggerHub.\nFor search queries all 3.0.x versions are grouped together\n"
        required: false
        type: "string"
        enum:
        - "2.0"
        - "3.0.0"
        - "3.0.1"
        - "3.0.2"
        - "3.0.3"
        x-exportParamName: "Oas"
        x-optionalDataType: "String"
      responses:
        "200":
          description: "a list of domains in APIs.json format. Each entry will be\
            \ listed with the 'X-Suggestion' property  which value will contain comma\
            \ separated list of available definitions matching the query"
          schema:
            $ref: "#/definitions/ApisJson"
      security:
      - TokenSecured: []
      - {}
      x-inflector-hidden: true
      x-swagger-router-controller: "com.smartbear.swaggerhub.registry.service.DomainsV1Service"
  /domains/{owner}/{domain}:
    get:
      tags:
      - "Domains"
      summary: "Retrieves an APIs.json listing for all domain versions for this owner\
        \ and domain"
      operationId: "getDomainVersions"
      parameters:
      - name: "owner"
        in: "path"
        description: "API or Domain owner identifier (case-sensitive)"
        required: true
        type: "string"
        x-exportParamName: "Owner"
      - name: "domain"
        in: "path"
        description: "domain identifier"
        required: true
        type: "string"
        x-exportParamName: "Domain"
      responses:
        "200":
          description: "a list of domain versions in APIs.json format"
          schema:
            $ref: "#/definitions/ApisJson"
      security:
      - TokenSecured: []
      - {}
      x-swagger-router-controller: "com.smartbear.swaggerhub.registry.service.DomainsV1Service"
    post:
      tags:
      - "Domains"
      summary: "Create or update a domain."
      operationId: "saveDomainDefinition"
      consumes:
      - "application/json"
      - "application/yaml"
      parameters:
      - name: "owner"
        in: "path"
        description: "API or Domain owner identifier (case-sensitive)"
        required: true
        type: "string"
        x-exportParamName: "Owner"
      - name: "domain"
        in: "path"
        description: "domain identifier"
        required: true
        type: "string"
        x-exportParamName: "Domain"
      - name: "isPrivate"
        in: "query"
        description: "Defines whether the API has to be private"
        required: false
        type: "boolean"
        default: false
        x-exportParamName: "IsPrivate"
        x-optionalDataType: "Bool"
      - name: "version"
        in: "query"
        description: "Domain version to create or update. If omitted, the version\
          \ will be taken from the `info.version` field in the definition."
        required: false
        type: "string"
        x-exportParamName: "Version"
        x-optionalDataType: "String"
      - in: "body"
        name: "definition"
        description: "the Swagger definition of this Domain"
        required: false
        schema:
          type: "string"
        x-exportParamName: "Definition"
        x-optionalDataType: "String"
      - name: "force"
        in: "query"
        description: "force update"
        required: false
        type: "boolean"
        x-exportParamName: "Force"
        x-optionalDataType: "Bool"
      - name: "projectName"
        in: "query"
        description: "The project to add the API or Domain to.\n"
        required: false
        type: "string"
        x-exportParamName: "ProjectName"
        x-optionalDataType: "String"
      responses:
        "200":
          description: "the domain was successfully saved"
        "201":
          description: "new domain was successfully saved"
        "205":
          description: "the domain was successfully saved and should be reloaded"
        "403":
          description: "maximum number of domains reached"
        "409":
          description: "can not overwrite a published domain version"
        "415":
          description: "invalid content type"
      security:
      - TokenSecured: []
      x-swagger-router-controller: "com.smartbear.swaggerhub.registry.service.DomainsV1Service"
    delete:
      tags:
      - "Domains"
      summary: "Deletes the specified domain"
      operationId: "deleteDomain"
      parameters:
      - name: "owner"
        in: "path"
        description: "API or Domain owner identifier (case-sensitive)"
        required: true
        type: "string"
        x-exportParamName: "Owner"
      - name: "domain"
        in: "path"
        description: "domain identifier"
        required: true
        type: "string"
        x-exportParamName: "Domain"
      - name: "force"
        in: "query"
        description: "force update"
        required: false
        type: "boolean"
        x-exportParamName: "Force"
        x-optionalDataType: "Bool"
      responses:
        "200":
          description: "the domain was successfully deleted"
        "403":
          description: "access denied"
        "404":
          description: "specified domain not found"
        "409":
          description: "the domain has published versions and can not be deleted"
        "424":
          description: "Can not delete because there are dependencies on this domain."
          schema:
            $ref: "#/definitions/Page"
      security:
      - TokenSecured: []
      x-swagger-router-controller: "com.smartbear.swaggerhub.registry.service.DomainsV1Service"
  /domains/{owner}/{domain}/.notifications:
    get:
      tags:
      - "Domains"
      summary: "Gets notification status for spec."
      operationId: "getDomainNotifications"
      parameters:
      - name: "owner"
        in: "path"
        description: "API or Domain owner identifier (case-sensitive)"
        required: true
        type: "string"
        x-exportParamName: "Owner"
      - name: "domain"
        in: "path"
        description: "domain identifier"
        required: true
        type: "string"
        x-exportParamName: "Domain"
      - name: "scope"
        in: "query"
        description: "Identifies requester"
        required: true
        type: "string"
        default: "USER"
        enum:
        - "USER"
        - "ALL"
        x-exportParamName: "Scope"
      - name: "version"
        in: "query"
        description: "version of the spec"
        required: false
        type: "string"
        x-exportParamName: "Version"
        x-optionalDataType: "String"
      - name: "filterViewers"
        in: "query"
        description: "Role as a criteria for filter list of notifiable users."
        required: false
        type: "boolean"
        x-exportParamName: "FilterViewers"
        x-optionalDataType: "Bool"
      responses:
        "200":
          description: "Separate lists of users with enabled and disabled notifications"
          schema:
            $ref: "#/definitions/NotificationPreferences"
        "403":
          description: "Access denied."
        "404":
          description: "Specified domain not found."
      security:
      - TokenSecured: []
      x-swagger-router-controller: "com.smartbear.swaggerhub.registry.service.DomainsV1Service"
    put:
      tags:
      - "Domains"
      summary: "Change notification status for spec for the user making the request"
      operationId: "setDomainNotifications"
      parameters:
      - name: "owner"
        in: "path"
        description: "API or Domain owner identifier (case-sensitive)"
        required: true
        type: "string"
        x-exportParamName: "Owner"
      - name: "domain"
        in: "path"
        description: "domain identifier"
        required: true
        type: "string"
        x-exportParamName: "Domain"
      responses:
        "201":
          description: "Setup notification status for given domain."
        "403":
          description: "Access denied."
        "404":
          description: "Specified API not found."
      security:
      - TokenSecured: []
      x-swagger-router-controller: "com.smartbear.swaggerhub.registry.service.DomainsV1Service"
    delete:
      tags:
      - "Domains"
      summary: "Disable notification status for spec"
      operationId: "disableDomainNotifications"
      parameters:
      - name: "owner"
        in: "path"
        description: "API or Domain owner identifier (case-sensitive)"
        required: true
        type: "string"
        x-exportParamName: "Owner"
      - name: "domain"
        in: "path"
        description: "domain identifier"
        required: true
        type: "string"
        x-exportParamName: "Domain"
      responses:
        "201":
          description: "Setup notification status for given domain."
        "403":
          description: "Access denied."
        "404":
          description: "Specified API not found."
      security:
      - TokenSecured: []
      x-swagger-router-controller: "com.smartbear.swaggerhub.registry.service.DomainsV1Service"
  /domains/{owner}/{domain}/rename:
    post:
      tags:
      - "Domains"
      summary: "Rename domain."
      operationId: "renameDomain"
      parameters:
      - name: "owner"
        in: "path"
        description: "API or Domain owner identifier (case-sensitive)"
        required: true
        type: "string"
        x-exportParamName: "Owner"
      - name: "domain"
        in: "path"
        description: "domain identifier"
        required: true
        type: "string"
        x-exportParamName: "Domain"
      - name: "newName"
        in: "query"
        description: "New name"
        required: true
        type: "string"
        maxLength: 60
        minLength: 3
        x-exportParamName: "NewName"
      - name: "force"
        in: "query"
        description: "force update"
        required: false
        type: "boolean"
        x-exportParamName: "Force"
        x-optionalDataType: "Bool"
      responses:
        "200":
          description: "Domain was successfully renamed."
        "403":
          description: "Access denied."
        "409":
          description: "API or Domain with the new name already exists."
        "424":
          description: "Can not rename because there are dependencies on this domain.\
            \ Use the `force` parameter to override this."
          schema:
            $ref: "#/definitions/Page"
      security:
      - TokenSecured: []
      x-swagger-router-controller: "com.smartbear.swaggerhub.registry.service.DomainsV1Service"
  /domains/{owner}/{domain}/.transfer:
    post:
      tags:
      - "Domains"
      summary: "transfers domain to another owner"
      description: ""
      operationId: "transferDomain"
      parameters:
      - name: "owner"
        in: "path"
        description: "API or Domain owner identifier (case-sensitive)"
        required: true
        type: "string"
        x-exportParamName: "Owner"
      - name: "domain"
        in: "path"
        description: "domain identifier"
        required: true
        type: "string"
        x-exportParamName: "Domain"
      - name: "newOwner"
        in: "query"
        description: "New owner"
        required: true
        type: "string"
        x-exportParamName: "NewOwner"
      - name: "force"
        in: "query"
        description: "force update"
        required: false
        type: "boolean"
        x-exportParamName: "Force"
        x-optionalDataType: "Bool"
      - name: "projectName"
        in: "query"
        description: "The project to add the API or Domain to.\n"
        required: false
        type: "string"
        x-exportParamName: "ProjectName"
        x-optionalDataType: "String"
      responses:
        "200":
          description: "the Swagger Domain was transfered"
          schema:
            type: "object"
      security:
      - TokenSecured: []
      x-inflector-hidden: true
      x-swagger-router-controller: "com.smartbear.swaggerhub.registry.service.DomainsV1Service"
  /domains/{owner}/{domain}/{version}:
    get:
      tags:
      - "Domains"
      summary: "Retrieves the Swagger definition for the specified domain and version"
      operationId: "getDomainDefinition"
      produces:
      - "application/json"
      - "application/yaml"
      parameters:
      - name: "owner"
        in: "path"
        description: "API or Domain owner identifier (case-sensitive)"
        required: true
        type: "string"
        x-exportParamName: "Owner"
      - name: "domain"
        in: "path"
        description: "domain identifier"
        required: true
        type: "string"
        x-exportParamName: "Domain"
      - name: "version"
        in: "path"
        description: "version identifier"
        required: true
        type: "string"
        x-exportParamName: "Version"
      responses:
        "200":
          description: "the Swagger domain in requested format"
          schema:
            type: "object"
        "404":
          description: "specified domain not found"
      security:
      - TokenSecured: []
      - {}
      x-swagger-router-controller: "com.smartbear.swaggerhub.registry.service.DomainsV1Service"
    delete:
      tags:
      - "Domains"
      summary: "Deletes a particular version of the specified domain"
      operationId: "deleteDomainVersion"
      parameters:
      - name: "owner"
        in: "path"
        description: "API or Domain owner identifier (case-sensitive)"
        required: true
        type: "string"
        x-exportParamName: "Owner"
      - name: "domain"
        in: "path"
        description: "domain identifier"
        required: true
        type: "string"
        x-exportParamName: "Domain"
      - name: "version"
        in: "path"
        description: "version identifier"
        required: true
        type: "string"
        x-exportParamName: "Version"
      - name: "force"
        in: "query"
        description: "force update"
        required: false
        type: "boolean"
        x-exportParamName: "Force"
        x-optionalDataType: "Bool"
      responses:
        "200":
          description: "the domain version was successfully deleted"
        "403":
          description: "access denied"
        "404":
          description: "specified domain not found"
        "409":
          description: "the domain version is published and can not be deleted or\
            \ it is the only version of this API"
        "424":
          description: "Can not delete because there are dependencies on this domain."
          schema:
            $ref: "#/definitions/Page"
      security:
      - TokenSecured: []
      x-swagger-router-controller: "com.smartbear.swaggerhub.registry.service.DomainsV1Service"
  /domains/{owner}/{domain}/{version}/fork:
    post:
      tags:
      - "Domains"
      summary: "Fork a domain."
      description: "Create a fork of the domain definition."
      operationId: "forkDomain"
      parameters:
      - name: "owner"
        in: "path"
        description: "API or Domain owner identifier (case-sensitive)"
        required: true
        type: "string"
        x-exportParamName: "Owner"
      - name: "domain"
        in: "path"
        description: "domain identifier"
        required: true
        type: "string"
        x-exportParamName: "Domain"
      - name: "version"
        in: "path"
        description: "version identifier"
        required: true
        type: "string"
        x-exportParamName: "Version"
      - in: "body"
        name: "forkVersion"
        description: "Fork version information."
        required: true
        schema:
          $ref: "#/definitions/ForkVersion"
        x-exportParamName: "ForkVersion"
      responses:
        "201":
          description: "Domain was successfully forked."
        "403":
          description: "Access denied."
        "409":
          description: "Conflict with existing existing definition."
      security:
      - TokenSecured: []
      x-swagger-router-controller: "com.smartbear.swaggerhub.registry.service.DomainsV1Service"
...
  /domains/{owner}/{domain}/{version}/clone:
    post:
      tags:
      - "Domains"
      summary: "Add new domain version."
      description: "Clone an existing domain version as a new version."
      operationId: "cloneDomain"
      parameters:
      - name: "owner"
        in: "path"
        description: "API or Domain owner identifier (case-sensitive)"
        required: true
        type: "string"
        x-exportParamName: "Owner"
      - name: "domain"
        in: "path"
        description: "domain identifier"
        required: true
        type: "string"
        x-exportParamName: "Domain"
      - name: "version"
        in: "path"
        description: "version identifier"
        required: true
        type: "string"
        x-exportParamName: "Version"
      - in: "body"
        name: "newVersion"
        description: "New domain version settings"
        required: true
        schema:
          $ref: "#/definitions/NewVersion"
        x-exportParamName: "NewVersion"
      responses:
        "201":
          description: "Domain version was successfully created"
        "400":
          description: "Invalid request"
        "403":
          description: "Access denied"
        "404":
          description: "Specified domain not found"
        "409":
          description: "Domain version already exists"
      security:
      - TokenSecured: []
      x-swagger-router-controller: "com.smartbear.swaggerhub.registry.service.DomainsV1Service"
...
securityDefinitions:
  TokenSecured:
    description: "Specify your SwaggerHub API key found in https://app.swaggerhub.com/settings/apiKey."
    type: "apiKey"
    name: "Authorization"
    in: "header"
definitions:
...
  ApisJson:
    type: "object"
    properties:
      name:
        type: "string"
      description:
        type: "string"
      url:
        type: "string"
      created:
        type: "string"
        format: "date-time"
      modified:
        type: "string"
        format: "date-time"
      specificationVersion:
        type: "string"
      offset:
        type: "integer"
        format: "int32"
      totalCount:
        type: "integer"
        format: "int64"
      apis:
        type: "array"
        items:
          $ref: "#/definitions/ApisJsonApi"
    example:
      offset: 0
      apis:
      - name: "name"
        description: "description"
        properties:
        - type: "type"
          value: "value"
          url: "url"
        - type: "type"
          value: "value"
          url: "url"
        tags:
        - "tags"
        - "tags"
      - name: "name"
        description: "description"
        properties:
        - type: "type"
          value: "value"
          url: "url"
        - type: "type"
          value: "value"
          url: "url"
        tags:
        - "tags"
        - "tags"
      created: "2000-01-23T04:56:07.000+00:00"
      name: "name"
      description: "description"
      modified: "2000-01-23T04:56:07.000+00:00"
      totalCount: 6
      url: "url"
      specificationVersion: "specificationVersion"
...
  NotificationPreferences:
    type: "object"
    properties:
      enabled:
        type: "array"
        items:
          type: "string"
    example:
      enabled:
      - "enabled"
      - "enabled"
...